+++ karma.conf.js
@@ -34,11 +34,12 @@ var firefoxFlags = {
     'media.navigator.permission.disabled': true,
     'media.navigator.streams.fake': true
 };
+var ci = process.env.TRAVIS || process.env.APPVEYOR;
 
 module.exports = function(config) {
     var configuration = {
         basePath: '',
-        frameworks: ['jasmine', 'jasmine-matchers'],
+        frameworks: ['jasmine', 'jasmine-matchers', 'detectBrowsers'],
         hostname: 'localhost',
         port: 9876,
         logLevel: config.LOG_INFO,
@@ -81,38 +82,70 @@ module.exports = function(config) {
             'karma-jasmine-matchers',
             'karma-chrome-launcher',
             'karma-firefox-launcher',
+            'karma-safari-launcher',
+            'karma-edge-launcher',
             'karma-coverage',
             'karma-coveralls',
-            'karma-verbose-reporter'
+            'karma-verbose-reporter',
+            'karma-detect-browsers'
         ],
-        browsers: ['Firefox_dev', 'Chrome_dev'],
+        detectBrowsers: {
+            enabled: true,
+            usePhantomJS: false,
+            preferHeadless: true,
+
+            postDetection: function(availableBrowsers) {
+                if (availableBrowsers.length > 1) {
+                    // use custom browser launchers
+                    var result = availableBrowsers;
+                    let cd = availableBrowsers.indexOf('ChromeHeadless');
+                    if (cd > -1) {
+                        availableBrowsers[cd] = 'Chrome_dev';
+                    }
+                    let fd = availableBrowsers.indexOf('FirefoxHeadless');
+                    if (fd > -1) {
+                        availableBrowsers[fd] = 'Firefox_dev';
+                    }
+                    let fh = availableBrowsers.indexOf('Firefox');
+                    if (fh > -1) {
+                        availableBrowsers[fh] = 'Firefox_dev';
+                    }
+                    let ch = availableBrowsers.indexOf('ChromiumHeadless');
+                    if (ch > -1) {
+                        availableBrowsers[ch] = 'Chromium_dev';
+                    }
+                    return result;
+                }
+            }
+        },
+        customLaunchers: {
+            Chrome_dev: {
+                base: 'ChromeHeadless',
+                flags: chromeFlags
+            },
+            Chromium_dev: {
+                base: 'ChromiumHeadless',
+                flags: chromeFlags
+            },
+            Firefox_dev: {
+                base: 'FirefoxHeadless',
+                prefs: firefoxFlags
+            }
+        },
         captureConsole: true,
         browserNoActivityTimeout: 50000,
         colors: true,
         reporters: ['verbose', 'progress', 'coverage'],
         coverageReporter: {
             type: 'html',
-            dir: 'coverage/'
+            dir: 'coverage'
         },
-        webpack: webpackConfig,
-        customLaunchers: {
-            Chrome_dev: {
-                base: 'Chrome',
-                flags: chromeFlags
-            },
-            Chrome_ci: {
-                base: 'ChromeHeadless',
-                flags: chromeFlags
-            },
-            Firefox_dev: {
-                base: 'Firefox',
-                prefs: firefoxFlags
-            }
-        }
+        webpack: webpackConfig
     };
 
-    if (process.env.TRAVIS || process.env.APPVEYOR) {
-        configuration.browsers = ['Chrome_ci'];
+    if (ci) {
+        configuration.browsers = ['Chrome_dev'];
+        configuration.detectBrowsers.enabled = false;
         configuration.singleRun = true;
 
         if (process.env.TRAVIS) {